chore: move config to outside utils
